AEDS
AEDS is a form application powered by Mircosoft SQL Server to provide
its database part.  AEDS allows users to buy and sell products, update
or remove products that they have already put for selling. Note that
both the products and users are stored in the database. Database allows
to store all the deals that have been done between buyers and sellers.
This data can be retrieved further in order to determine the most active
user or the most favourite product.